requirements.txt
altgraph==0.17.4
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==24.0
pefile==2023.2.7
pillow==10.3.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.5
python-dotenv==1.0.0
pywin32-ctypes==0.2.2
redis==5.0.1
requests==2.31.0
setuptools==69.5.1
six==1.16.0
spotipy==2.23.0
sv-ttk==2.6.0
tk==0.1.0
tkinter-tooltip==3.0.0
urllib3==2.1.0
Werkzeug==3.0.1
